🏗️ structure: 클래스 테스트 파일 분리
